Fix tests fix assembly report
