+++ eva_sub_cli/__init__.py
@@ -16,12 +16,3 @@ SUBMISSION_WS_VAR = 'SUBMISSION_WS_URL'
 ENA_WEBIN_ACCOUNT_VAR = 'ENA_WEBIN_ACCOUNT'
 ENA_WEBIN_PASSWORD_VAR = 'ENA_WEBIN_PASSWORD'
 
-
-def is_submission_dir_writable(submission_dir):
-    if not os.path.exists(submission_dir):
-        os.makedirs(submission_dir)
-    if not os.path.isdir(submission_dir):
-        return False
-    if not os.access(submission_dir, os.W_OK):
-        return False
-    return True
