+++ tests/test_submit.py
@@ -6,7 +6,8 @@ from unittest.mock import MagicMock, patch, Mock
 import yaml
 from ebi_eva_common_pyutils.config import WritableConfig
 
-from eva_sub_cli import SUB_CLI_CONFIG_FILE, is_submission_dir_writable
+from eva_sub_cli import SUB_CLI_CONFIG_FILE
+from eva_sub_cli.utils import is_submission_dir_writable
 from eva_sub_cli.validators.validator import READY_FOR_SUBMISSION_TO_EVA
 from eva_sub_cli.submit import StudySubmitter, SUB_CLI_CONFIG_KEY_SUBMISSION_ID, SUB_CLI_CONFIG_KEY_SUBMISSION_UPLOAD_URL
